Create a new car model for SMCE
Merged the RayCar with a opensource godot model. Uses all the same
attachments that RayCar uses, this means that to run the new model you
need access to the uncompiled project as far as I am aware. duplicated
the code for the car and spring however into our new model so that we
can tweak values unrelated to the sensors and attachments.