New Race Map with Bots
This new smce mod includes a new race track with cars that drive
themselves. booster plates to increase the speed of smce during the
race. a minimap to keep track of the posisition of everyone. a clock
that keeps track of the time of the race and each cars finish time.